HDDS-11031. Merge BaseApprover abstract class with the DefaultApprover.
